refactor(路由): 重构路由组件
