+++ src/App.vue
@@ -13,6 +13,7 @@ import zh_CN from 'ant-design-vue/lib/locale-provider/zh_CN';
 import en_US from 'ant-design-vue/lib/locale-provider/en_US';
 
 export default {
+  name: 'App',
   data() {
     return {
       lang: {
