--- src/components/index/WriteArticleIndex.vue
@@ -1,19 +1,0 @@
-<template>
-  <div id="write-article-index">
-    <write-article />
-  </div>
-</template>
-
-<script>
-import WriteArticle from '@/components/article/WriteArticle';
-
-export default {
-  components: { WriteArticle },
-};
-</script>
-
-<style>
-#write-article-index {
-  height: 100%;
-}
-</style>
