+++ src/views/book/index.vue
@@ -1,0 +1,12 @@
+<template>
+  <nan-sheng-value />
+</template>
+
+<script>
+import NanShengValue from './components/NanShengValue';
+
+export default {
+  name: '',
+  components: { NanShengValue },
+};
+</script>
