+++ src/views/home/index.vue
@@ -331,27 +331,3 @@ export default {
   },
 };
 </script>
-
-<style>
-.index-drawer-wrap .ant-drawer-content-wrapper {
-  width: 250px !important;
-}
-
-#components-layout-demo-custom-trigger .trigger {
-  font-size: 18px;
-  line-height: 64px;
-  padding: 0 24px;
-  cursor: pointer;
-  transition: color 0.3s;
-}
-
-#components-layout-demo-custom-trigger .trigger:hover {
-  color: #1890ff;
-}
-
-#components-layout-demo-custom-trigger .logo {
-  height: 64px;
-  background: rgba(255, 255, 255, 0.2);
-  margin: 0;
-}
-</style>
