+++ src/views/settings/account/components/BindEmail.vue
