+++ src/views/settings/account/components/BindPhone.vue
