+++ src/views/settings/account/components/ChangePassword.vue
